requirements.txt
pymysql
pandas
kafka-python
pyspark
